Fix a position field bug on the controller when saving an ordination
This commit fixes a bug that happens when one tries to use a fully
qualified column name for the position column, such as the string
"my_table.position".

I needed to use a fully qualified name in my config to avoid ambiguities
in the column names, because one of the tables of the query was already
referencing a "position" field.

The controller code that saves the list ordination sends a message to
the ActiveRecord object, in order to write a number to the position
field:

model.send("#{@position_field}=".to_sym, (key.to_i + 1))

That was blowing up, because it tries to send a message named
"my_table.position=" to the ActiveRecord object, when the actual and
correct message is "position=".