Create a new object_label option to customize the object label
I needed the object_label of each list item to be different from
the RailsAdmin default object_label.

Reason: the RailsAdmin "list" GET method can show a table with many
columns, which may have valuable context and information (specially in
complex models). We don't have additional columns in the ordination
list; we only have the object label.

So I implemented an object_label option to provide more flexibility in
what to show for each item. It works like this:

- You can specify an object_label: a Symbol or a Proc

- If you don't, it fallbacks to the previous implemented method, which
  uses RailsAdmin object_label.